refactor(tests): build job item moved to common
